code climate: refactor nfsession
